refactor: reorganize binary helpers and mod config
