+++ uplink/src/base/bridge/streams.rs
@@ -6,7 +6,7 @@ use log::{error, info, trace};
 
 use super::stream::{self, StreamStatus};
 use super::{Point, StreamMetrics};
-use crate::base::StreamConfig;
+use crate::config::StreamConfig;
 use crate::{Config, Package, Stream};
 
 use super::delaymap::DelayMap;
