+++ uplink/src/collector/installer.rs
@@ -6,8 +6,7 @@ use tar::Archive;
 use tokio::process::Command;
 
 use super::downloader::DownloadFile;
-use crate::base::{bridge::BridgeTx, InstallerConfig};
-use crate::{Action, ActionResponse};
+use crate::{base::bridge::BridgeTx, config::InstallerConfig, Action, ActionResponse};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
