+++ uplink/src/lib.rs
@@ -66,213 +66,15 @@ use log::error;
 
 pub mod base;
 pub mod collector;
-
-pub mod config {
-    use crate::base::{bridge::stream::MAX_BUFFER_SIZE, StreamConfig};
-    pub use crate::base::{Config, Persistence, Stats};
-    use config::{Environment, File, FileFormat};
-    use std::fs;
-    use structopt::StructOpt;
-
-    #[derive(StructOpt, Debug)]
-    #[structopt(name = "uplink", about = "collect, batch, compress, publish")]
-    pub struct CommandLine {
-        /// Binary version
-        #[structopt(skip = env ! ("VERGEN_BUILD_SEMVER"))]
-        pub version: String,
-        /// Build profile
-        #[structopt(skip = env ! ("VERGEN_CARGO_PROFILE"))]
-        pub profile: String,
-        /// Commit SHA
-        #[structopt(skip = env ! ("VERGEN_GIT_SHA"))]
-        pub commit_sha: String,
-        /// Commit SHA
-        #[structopt(skip = env ! ("VERGEN_GIT_COMMIT_TIMESTAMP"))]
-        pub commit_date: String,
-        /// config file
-        #[structopt(short = "c", help = "Config file")]
-        pub config: Option<String>,
-        /// config file
-        #[structopt(short = "a", help = "Authentication file")]
-        pub auth: String,
-        /// log level (v: info, vv: debug, vvv: trace)
-        #[structopt(short = "v", long = "verbose", parse(from_occurrences))]
-        pub verbose: u8,
-        /// list of modules to log
-        #[structopt(short = "m", long = "modules")]
-        pub modules: Vec<String>,
-    }
-
-    const DEFAULT_CONFIG: &str = r#"    
-    [mqtt]
-    max_packet_size = 256000
-    max_inflight = 100
-    keep_alive = 30
-    network_timeout = 30
-
-    [stream_metrics]
-    enabled = false
-    bridge_topic = "/tenants/{tenant_id}/devices/{device_id}/events/uplink_stream_metrics/jsonarray"
-    serializer_topic = "/tenants/{tenant_id}/devices/{device_id}/events/uplink_serializer_stream_metrics/jsonarray"
-    blacklist = []
-    timeout = 10
-
-    [serializer_metrics]
-    enabled = false
-    topic = "/tenants/{tenant_id}/devices/{device_id}/events/uplink_serializer_metrics/jsonarray"
-    timeout = 10
-
-    [mqtt_metrics]
-    enabled = true
-    topic = "/tenants/{tenant_id}/devices/{device_id}/events/uplink_mqtt_metrics/jsonarray"
-
-    [action_status]
-    topic = "/tenants/{tenant_id}/devices/{device_id}/action/status"
-    buf_size = 1
-    flush_period = 2
-    priority = 255 # highest priority for quick delivery of action status info to platform
-
-    [streams.device_shadow]
-    topic = "/tenants/{tenant_id}/devices/{device_id}/events/device_shadow/jsonarray"
-    flush_period = 5
-
-    [streams.logs]
-    topic = "/tenants/{tenant_id}/devices/{device_id}/events/logs/jsonarray"
-    buf_size = 32
-
-    [system_stats]
-    enabled = true
-    process_names = ["uplink"]
-    update_period = 30
-"#;
-
-    /// Reads config file to generate config struct and replaces places holders
-    /// like bike id and data version
-    pub fn initialize(auth_config: &str, uplink_config: &str) -> Result<Config, anyhow::Error> {
-        let config = config::Config::builder()
-            .add_source(File::from_str(DEFAULT_CONFIG, FileFormat::Toml))
-            .add_source(File::from_str(uplink_config, FileFormat::Toml))
-            .add_source(File::from_str(auth_config, FileFormat::Json))
-            .add_source(Environment::default())
-            .build()?;
-
-        let mut config: Config = config.try_deserialize()?;
-
-        // Create directory at persistence_path if it doesn't already exist
-        fs::create_dir_all(&config.persistence_path).map_err(|_| {
-            anyhow::Error::msg(format!(
-                "Permission denied for creating persistence directory at \"{}\"",
-                config.persistence_path.display()
-            ))
-        })?;
-
-        // replace placeholders with device/tenant ID
-        let tenant_id = config.project_id.trim();
-        let device_id = config.device_id.trim();
-        for config in config.streams.values_mut() {
-            replace_topic_placeholders(&mut config.topic, tenant_id, device_id);
-        }
-
-        replace_topic_placeholders(&mut config.action_status.topic, tenant_id, device_id);
-        replace_topic_placeholders(&mut config.stream_metrics.bridge_topic, tenant_id, device_id);
-        replace_topic_placeholders(
-            &mut config.stream_metrics.serializer_topic,
-            tenant_id,
-            device_id,
-        );
-        replace_topic_placeholders(&mut config.serializer_metrics.topic, tenant_id, device_id);
-        replace_topic_placeholders(&mut config.mqtt_metrics.topic, tenant_id, device_id);
-
-        // for config in [&mut config.serializer_metrics, &mut config.stream_metrics] {
-        //     if let Some(topic) = &config.topic {
-        //         let topic = topic.replace("{tenant_id}", tenant_id);
-        //         let topic = topic.replace("{device_id}", device_id);
-        //         config.topic = Some(topic);
-        //     }
-        // }
-
-        if config.system_stats.enabled {
-            for stream_name in [
-                "uplink_disk_stats",
-                "uplink_network_stats",
-                "uplink_processor_stats",
-                "uplink_process_stats",
-                "uplink_component_stats",
-                "uplink_system_stats",
-            ] {
-                config.stream_metrics.blacklist.push(stream_name.to_owned());
-                let stream_config = StreamConfig {
-                    topic: format!(
-                        "/tenants/{tenant_id}/devices/{device_id}/events/{stream_name}/jsonarray"
-                    ),
-                    buf_size: config.system_stats.stream_size.unwrap_or(MAX_BUFFER_SIZE),
-                    ..Default::default()
-                };
-                config.streams.insert(stream_name.to_owned(), stream_config);
-            }
-        }
-
-        #[cfg(any(target_os = "linux", target_os = "android"))]
-        if let Some(buf_size) = config.logging.as_ref().and_then(|c| c.stream_size) {
-            let stream_config =
-                config.streams.entry("logs".to_string()).or_insert_with(|| StreamConfig {
-                    topic: format!(
-                        "/tenants/{tenant_id}/devices/{device_id}/events/logs/jsonarray"
-                    ),
-                    buf_size: 32,
-                    ..Default::default()
-                });
-            stream_config.buf_size = buf_size;
-        }
-
-        let action_topic_template = "/tenants/{tenant_id}/devices/{device_id}/actions";
-        let mut device_action_topic = action_topic_template.to_string();
-        replace_topic_placeholders(&mut device_action_topic, tenant_id, device_id);
-        config.actions_subscription = device_action_topic;
-
-        Ok(config)
-    }
-
-    // Replace placeholders in topic strings with configured values for tenant_id and device_id
-    fn replace_topic_placeholders(topic: &mut String, tenant_id: &str, device_id: &str) {
-        *topic = topic.replace("{tenant_id}", tenant_id);
-        *topic = topic.replace("{device_id}", device_id);
-    }
-
-    #[derive(Debug, thiserror::Error)]
-    pub enum ReadFileError {
-        #[error("Auth file not found at {0}")]
-        Auth(String),
-        #[error("Config file not found at {0}")]
-        Config(String),
-    }
-
-    fn read_file_contents(path: &str) -> Option<String> {
-        fs::read_to_string(path).ok()
-    }
-
-    pub fn get_configs(
-        commandline: &CommandLine,
-    ) -> Result<(String, Option<String>), ReadFileError> {
-        let auth = read_file_contents(&commandline.auth)
-            .ok_or_else(|| ReadFileError::Auth(commandline.auth.to_string()))?;
-        let config = match &commandline.config {
-            Some(path) => Some(
-                read_file_contents(path).ok_or_else(|| ReadFileError::Config(path.to_string()))?,
-            ),
-            None => None,
-        };
-
-        Ok((auth, config))
-    }
-}
+pub mod config;
 
 pub use base::actions::{Action, ActionResponse};
 use base::bridge::{Bridge, Package, Payload, Point, StreamMetrics};
 use base::mqtt::Mqtt;
 use base::serializer::{Serializer, SerializerMetrics};
-pub use base::{ActionRoute, Config};
 pub use collector::{simulator, tcpjson::TcpJson};
+use config::ActionRoute;
+pub use config::Config;
 pub use storage::Storage;
 
 /// Spawn a named thread to run the function f on
